Languages Body moved to its own component.
