+++ src/App.js
@@ -1,6 +1,6 @@
 
 import './App.css';
-import LanguagesBody from './components/LangBody/LanguagesBody';
+import LanguagesBody from './components/Body/Languages/LanguagesBody';
 import MainNavbar from './components/Navbar/MainNavbar';
 
 function App() {
