+++ src/components/Body/Languages/LanguagesBody.css
@@ -1,0 +1,6 @@
+@media screen and (min-width:1024px) {
+    .grid {
+        display: grid;
+        grid-template-columns: 30% 70%;
+    }
+}
