+++ src/components/Body/Languages/LanguagesContent/LanguagesContent.css
@@ -3,24 +3,19 @@
     outline: 1px solid red;
 } */
 
-.container{
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    background-color: white;
-}
 
 icon{
     font-size: 1.5rem;
     padding-right: 1.2rem;
 }
 
-.lang_body{
+.lang_content{
     display: flex;
     flex-direction: column;
     justify-content: center;
     padding: 4%;
     font-size: 1rem;
+    background-color: white;
 }
 
 .lang_stars{
@@ -51,12 +46,7 @@ li span {
   }
 
 @media screen and (min-width:1024px) {
-    .grid {
-        display: grid;
-        grid-template-columns: 30% 70%;
-    }
-
-    .lang_body{
+    .lang_content{
         background-color: white;
         border-left: 100px;
         border-color: lightgrey;
